+++ src/Spreads.Collections/Collections/IndexedMap.fs
@@ -227,15 +227,15 @@ type internal IndexedMap<'K,'V when 'K : comparison>
     member this.AddFirst(k, v) = this.AddFirst(k, v)
     member this.Remove(k) = this.Remove(k)
     member this.RemoveFirst([<Out>] result: byref<KeyValuePair<'K, 'V>>) = 
-      this.RemoveFirst(&result) |> ignore
+      this.RemoveFirst(&result)
     member this.RemoveLast([<Out>] result: byref<KeyValuePair<'K, 'V>>) = 
-      this.RemoveLast(&result) |> ignore
+      this.RemoveLast(&result)
     member this.RemoveMany(key:'K,direction:Lookup) = 
-      this.RemoveMany(key, direction) |> ignore
-      ()
-    member this.Append(appendMap:IReadOnlyOrderedMap<'K,'V>) =
-      for i in appendMap do
-        this.AddLast(i.Key, i.Value)
+      this.RemoveMany(key, direction)
+    member this.Append(appendMap:IReadOnlyOrderedMap<'K,'V>, option:AppendOption) =
+//      for i in appendMap do
+//        this.AddLast(i.Key, i.Value)
+      raise (NotImplementedException("TODO append impl"))
     
 
   interface IImmutableOrderedMap<'K,'V> with